Updated heuristics and weights to use doubles not floats. This allows for better stream support.

+++ program/src/main/java/uk/ac/rhul/project/expectimax/LeafNodeBehaviour.java
@@ -18,7 +18,7 @@ public class LeafNodeBehaviour implements NodeBehaviour
     }
 
     @Override
-    public float applyHeuristic(Heuristic heuristic)
+    public double applyHeuristic(Heuristic heuristic)
     {
         return state.applyHeuristic(heuristic);
     }
