+++ program/src/main/java/uk/ac/rhul/project/heursitics/Heuristic.java
@@ -12,7 +12,7 @@ public interface Heuristic
      * @param state The game state to be evaluated.
      * @return A float value representing how good the state is.
      */
-    float heuristic(GameState state);
+    double heuristic(GameState state);
 
     /**
      * Returns the name of heuristic function
